// Semana10_angular/project-nodejs/src/models/product.model.ts
// export interface Product {
//     id: number;
//     descricao: string;
//     imagem: string;
//     preco: number;
//     quantidade: number;
//     disponivel: boolean;
// }

export interface Colaborator {
  id: number;
  name: string;
  wage: number;
  ocupation: string;
  openToWork: boolean;
}